Separate and import AMP globals
