+++ src/core/data-structures/amp-globals.d.ts
@@ -1,0 +1,8 @@
+export {};
+
+declare global {
+  interface Window {
+    // Counter for the DomBaseWeakRef polyfill.
+    __AMP_WEAKREF_ID?: number;
+  }
+}
