+++ src/core/data-structures/globals.d.ts
@@ -1,10 +1,7 @@
-export {};
+export * from './amp-globals.d';
 
 declare global {
   interface Window {
     WeakRef?: typeof WeakRef;
-
-    // Counter for the DomBaseWeakRef polyfill.
-    __AMP_WEAKREF_ID?: number;
   }
 }
