+++ src/core/globals.d.ts
@@ -1,29 +1,8 @@
-export {};
+export * from './amp-globals.d';
 
 declare global {
   interface Window {
-    // Global property set by test some harnesses to signal a testing environment.
-    __AMP_TEST?: boolean;
-
     // Global property set by test some harnesses to signal karma testing environment.
     __karma__?: boolean;
-
-
-    // AMP Mode, used to force an override in tests.
-    __AMP_MODE: AmpModeObject;
-
-    // AMP Runtime settings, configuration, and environment/build constants.
-    AMP_CONFIG?: AmpConfig & {[key: string]: (boolean|string)};
-  }
-
-  interface AmpConfig {
-    test?: boolean;
-    localDev?: boolean;
-    canary?: boolean;
-    type?: string;
-  }
-
-  interface AmpModeObject {
-    esm: boolean;
   }
 }
