+++ src/core/types/globals.d.ts
