+++ src/core/types/string/amp-globals.d.ts
@@ -1,0 +1,8 @@
+export {};
+
+declare global {
+  interface Location {
+    // Set by a viewer when it removes the fragment.
+    originalHash?: string;
+  }
+}
