+++ src/core/types/string/globals.d.ts
@@ -1,12 +1,7 @@
-export {};
+export * from './amp-globals.d';
 
 declare global {
   interface Window {
     msCrypto?: Crypto;
-  }
-
-  interface Location {
-    // Set by a viewer when it removes the fragment.
-    originalHash?: string;
   }
 }
